pyproject.toml
[project]
name = "ensure-project-acl"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pyyaml"
]

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
]
